requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.5
requests==2.26.0
beautifulsoup4==4.10.0
lxml==4.6.3
pytest==7.1.2
flask-cors==3.0.10
werkzeug==2.2.2
python-dotenv==0.19.1
flask-session==0.4.0
selenium==4.18.1
itsdangerous==2.0.1